Add fxlogger, level, go.mod, go.sum, and update .gitignore
Added new SlogLogger file "fxlogger.go" to the otris package, which serves as a Fx event logger utilizing the slog library. A new "level.go" was also added to specify a custom logging level for FxHandler. Created the "go.mod" and "go.sum" files which specify the required versions of Go and the required libraries. The .gitignore file was updated to ignore Goland workspace files.

Signed-off-by: Timur Kulakov <kulakovcore@gmail.com>